Update the styles to better cater to larger resolutions
Okay, that's not all of what's happening here, but it's the main
inspiration. I looked at the page on a large display, by accident, and
saw that it was massive. Given that, I added this change to add a
meaningful width limit to it. I also updated the links to be a bit
snazier too.