Add ability to differentiate between environments
This change does so by using PHP Dotenv, which sets a MODE variable
which the code can read, to determine which mode its in. What it also
does is to set the posts data file directory based on the environment as
well, which makes the application testable, as the test files can be set
and not be subject to change.