Add caching support
This change adds laminas-cache and
laminas-cache-storage-adapter-memcached as required packages to provide
caching support to the application.

There are a host of caching libraries available in PHP
(https://packagist.org/?query=cache), so why these two? Honestly, I'm
most familiar with working with Laminas packages, so they fit like the
proverbial glove. Also, they just slot into place, without requiring a
particular framework to be used or for an application to work in a
specific way. Given that, they're my choice!